+++ cmd/plugins/juju-metadata/validateimagemetadata_test.go
@@ -153,7 +153,7 @@ func (s *ValidateImageMetadataSuite) TestEc2LocalMetadataUsingIncompleteEnvironm
 	c.Assert(code, gc.Equals, 1)
 	errOut := ctx.Stderr.(*bytes.Buffer).String()
 	strippedOut := strings.Replace(errOut, "\n", "", -1)
-	c.Check(strippedOut, gc.Matches, `error: environment has no access-key or secret-key`)
+	c.Check(strippedOut, gc.Matches, `error: .*environment has no access-key or secret-key`)
 }
 
 func (s *ValidateImageMetadataSuite) TestEc2LocalMetadataWithManualParams(c *gc.C) {
