+++ provider/ec2/environ_test.go
